examples/python-xiazai/requirements.txt
mysql-connector-python
pymongo
qrcode[pil]
pyqrcode
